services/anomaly/requirements.txt
fastapi
gunicorn
uvicorn[standard]
google-cloud-bigquery
google-cloud-pubsub

